Move view create entry point to a controller
